Rename files to remove `reactive-*`
